try to maximize coverage
