+++ goversioninfo_test.go
@@ -7,10 +7,13 @@ package goversioninfo
 import (
 	"bytes"
 	"fmt"
+	"io"
 	"io/ioutil"
 	"os"
 	"path/filepath"
 	"testing"
+
+	"github.com/akavel/rsrc/coff"
 )
 
 // *****************************************************************************
@@ -367,19 +370,25 @@ func TestStr2Uint32(t *testing.T) {
 	for _, tt := range []struct {
 		in  string
 		out uint32
-	}{{"0", 0}, {"", 0}, {"FFEF", 65519}} {
+	}{{"0", 0}, {"", 0}, {"FFEF", 65519}, {"\x00\x00", 0}} {
 		got := str2Uint32(tt.in)
 		if got != tt.out {
 			t.Errorf("%q: awaited %d, got %d.", tt.in, tt.out, got)
 		}
 	}
 }
+
+var unmarshals = []struct {
+	in      string
+	needErr bool
+}{
+	{"", false}, {"A", true}, {"1", false}, {`"FfeF"`, false},
+	{`"FfeF`, true}, {`"FXXX"`, true},
+}
+
 func TestLangID(t *testing.T) {
 	var lng LangID
-	for _, tt := range []struct {
-		in      string
-		needErr bool
-	}{{"", false}, {"A", true}, {"1", false}, {`"FfeF"`, false}} {
+	for _, tt := range unmarshals {
 		if err := lng.UnmarshalJSON([]byte(tt.in)); tt.needErr && err == nil {
 			t.Errorf("%q: needed error, got nil.", tt.in)
 		} else if !tt.needErr && err != nil {
@@ -390,14 +399,46 @@ func TestLangID(t *testing.T) {
 
 func TestCharsetID(t *testing.T) {
 	var cs CharsetID
-	for _, tt := range []struct {
-		in      string
-		needErr bool
-	}{{"", false}, {"A", true}, {"1", false}, {`"FfeF"`, false}} {
+	for _, tt := range unmarshals {
 		if err := cs.UnmarshalJSON([]byte(tt.in)); tt.needErr && err == nil {
 			t.Errorf("%q: needed error, got nil.", tt.in)
 		} else if !tt.needErr && err != nil {
 			t.Errorf("%q: got error: %v", tt.in, err)
 		}
 	}
+}
+
+func TestWriteCoff(t *testing.T) {
+	tempFh, err := ioutil.TempFile("", "goversioninfo-test-")
+	if err != nil {
+		t.Fatalf("temp file: %v", err)
+	}
+	tempfn := tempFh.Name()
+	tempFh.Close()
+	defer os.Remove(tempfn)
+
+	if err := writeCoff(nil, ""); err == nil {
+		t.Errorf("needed error, got nil")
+	}
+	if err := writeCoff(nil, tempfn); err != nil {
+		t.Errorf("got %v", err)
+	}
+
+	if err := writeCoffTo(badWriter{writeErr: io.EOF}, coff.NewRSRC()); err == nil {
+		t.Errorf("needed write error, got nil")
+	}
+	if err := writeCoffTo(badWriter{closeErr: io.EOF}, nil); err == nil {
+		t.Errorf("needed close error, got nil")
+	}
+}
+
+type badWriter struct {
+	writeErr, closeErr error
+}
+
+func (w badWriter) Write(p []byte) (int, error) {
+	return len(p), w.writeErr
+}
+func (w badWriter) Close() error {
+	return w.closeErr
 }
